+++ fabric-biome-api-v1-bridge/src/main/java/net/fabricmc/fabric/api/biome/v1/ModificationPhase.java
@@ -1,0 +1,8 @@
+package net.fabricmc.fabric.api.biome.v1;
+
+public enum ModificationPhase {
+    ADDITIONS,
+    REMOVALS,
+    REPLACEMENTS,
+    POST_PROCESSING
+}